Version 1.9.0

+++ service/src/main/java/org/grnet/cat/services/registry/CriterionService.java
@@ -4,8 +4,10 @@ import io.quarkus.hibernate.orm.panache.Panache;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.UriInfo;
+import org.apache.commons.lang3.StringUtils;
 import org.grnet.cat.dtos.registry.criterion.CriterionRequest;
 import org.grnet.cat.dtos.registry.criterion.CriterionResponse;
 import org.grnet.cat.dtos.registry.criterion.CriterionUpdate;
@@ -15,19 +17,17 @@ import org.grnet.cat.dtos.registry.criterion.PrincipleCriterionResponse;
 import org.grnet.cat.dtos.registry.template.MetricNode;
 import org.grnet.cat.dtos.registry.template.Node;
 import org.grnet.cat.dtos.registry.template.TestNode;
-import org.grnet.cat.entities.registry.Imperative;
-import org.grnet.cat.entities.registry.TypeCriterion;
+import org.grnet.cat.utils.TestParamsTransformer;
+import org.grnet.cat.entities.registry.*;
 import org.grnet.cat.exceptions.UniqueConstraintViolationException;
 import org.grnet.cat.mappers.registry.CriteriaMapper;
+import org.grnet.cat.mappers.registry.MotivationMapper;
 import org.grnet.cat.mappers.registry.PrincipleCriterionMapper;
-import org.grnet.cat.repositories.registry.CriterionMetricRepository;
-import org.grnet.cat.repositories.registry.CriterionRepository;
-import org.grnet.cat.repositories.registry.ImperativeRepository;
-import org.grnet.cat.repositories.registry.TypeCriterionRepository;
+import org.grnet.cat.repositories.registry.*;
 import org.jboss.logging.Logger;
 
-import java.util.HashMap;
-import java.util.Objects;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @ApplicationScoped
 public class CriterionService {
@@ -44,6 +44,11 @@ public class CriterionService {
     @Inject
     CriterionMetricRepository criterionMetricRepository;
 
+    @Inject
+    PrincipleCriterionRepository principleCriterionRepository;
+    @Inject
+    CriterionActorRepository criterionActorRepository;
+
     private static final Logger LOG = Logger.getLogger(CriterionService.class);
 
     /**
@@ -54,10 +59,14 @@ public class CriterionService {
      * @param uriInfo The Uri Info.
      * @return A PageResource containing the criteria items in the requested page.
      */
-    public PageResource<CriterionResponse> listAll(int page, int size, UriInfo uriInfo) {
+    public PageResource<CriterionResponse> listAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {
 
-        var criteriaPage = criteriaRepository.fetchCriteriaByPage(page, size);
-        var criteriaDTOs = CriteriaMapper.INSTANCE.criteriaToDtos(criteriaPage.list());
+        var criteriaPage = criteriaRepository.fetchCriteriaByPage(search, sort, order, page, size);
+
+        var criteriaDTOs = criteriaPage.list().stream()
+                .map(this::criterionResponseWithMotivations)
+                .collect(Collectors.toList());
+
         return new PageResource<>(criteriaPage, criteriaDTOs, uriInfo);
     }
 
@@ -69,15 +78,16 @@ public class CriterionService {
      */
     public CriterionResponse findById(String id) {
 
-        var criteria = criteriaRepository.findById(id);
-        return CriteriaMapper.INSTANCE.criteriaToDto(criteria);
+        var criterion = criteriaRepository.findById(id);
+
+        return criterionResponseWithMotivations(criterion);
     }
 
     /**
      * Creates a new criteria item.
      *
      * @param criteriaRequestDto The criteria item to create.
-     * @param userId The ID of the user creating the criteria.
+     * @param userId             The ID of the user creating the criteria.
      * @return The created criteria item.
      */
     @Transactional
@@ -86,7 +96,7 @@ public class CriterionService {
         if (criteriaRepository.notUnique("cri", criteriaRequestDto.cri.toUpperCase())) {
             throw new UniqueConstraintViolationException("cri", criteriaRequestDto.cri.toUpperCase());
         }
-        
+
         var criteria = CriteriaMapper.INSTANCE.criteriaToEntity(criteriaRequestDto);
 
         criteria.setPopulatedBy(userId);
@@ -101,31 +111,42 @@ public class CriterionService {
     /**
      * Updates an existing criteria item.
      *
-     * @param id                  The ID of the criteria item to update.
-     * @param criteriaUpdateDto  The updated criteria item.
-     * @param userId              The ID of the user updating the criteria.
+     * @param id                The ID of the criteria item to update.
+     * @param criteriaUpdateDto The updated criteria item.
+     * @param userId            The ID of the user updating the criteria.
      * @return The updated criteria item.
      */
+    //@CheckPublishedRelation(permittedStatus = false,type = PublishEntityType.CRITERION)  // Only proceed if `published` is false
     @Transactional
     public CriterionResponse update(String id, CriterionUpdate criteriaUpdateDto, String userId) {
 
-        if (!Objects.isNull(criteriaUpdateDto.cri) && criteriaRepository.notUnique("cri", criteriaUpdateDto.cri.toUpperCase())) {
-            throw new UniqueConstraintViolationException("cri", criteriaUpdateDto.cri.toUpperCase());
+        if(criterionActorRepository.existCriterionInStatus(id, Boolean.TRUE)){
+            throw new ForbiddenException("No action permitted , criterion exists in a published motivation");
         }
 
         var criteria = criteriaRepository.findById(id);
 
+        var currentCri = criteria.getCri();
+        var updateCri = StringUtils.isNotEmpty(criteriaUpdateDto.cri) ? criteriaUpdateDto.cri.toUpperCase() : currentCri;
+
+        if (StringUtils.isNotEmpty(updateCri) && !updateCri.equals(currentCri)) {
+
+            if (criteriaRepository.notUnique("cri", updateCri)) {
+                throw new UniqueConstraintViolationException("cri", updateCri);
+            }
+        }
+
         CriteriaMapper.INSTANCE.updateCriteria(criteriaUpdateDto, criteria);
 
-        if(!Objects.isNull(criteriaUpdateDto.imperative)){
+        if (!Objects.isNull(criteriaUpdateDto.imperative)) {
 
-            imperativeRepository.findByIdOptional(criteriaUpdateDto.imperative).orElseThrow(()-> new NotFoundException("There is no Imperative Type with the following id: "+criteriaUpdateDto.imperative));
+            imperativeRepository.findByIdOptional(criteriaUpdateDto.imperative).orElseThrow(() -> new NotFoundException("There is no Imperative Type with the following id: " + criteriaUpdateDto.imperative));
             criteria.setImperative(Panache.getEntityManager().getReference(Imperative.class, criteriaUpdateDto.imperative));
         }
 
-        if(!Objects.isNull(criteriaUpdateDto.typeCriterion)){
+        if (!Objects.isNull(criteriaUpdateDto.typeCriterion)) {
 
-            typeCriterionRepository.findByIdOptional(criteriaUpdateDto.typeCriterion).orElseThrow(()-> new NotFoundException("There is no Type Criterion with the following id: "+criteriaUpdateDto.typeCriterion));
+            typeCriterionRepository.findByIdOptional(criteriaUpdateDto.typeCriterion).orElseThrow(() -> new NotFoundException("There is no Type Criterion with the following id: " + criteriaUpdateDto.typeCriterion));
             criteria.setTypeCriterion(Panache.getEntityManager().getReference(TypeCriterion.class, criteriaUpdateDto.typeCriterion));
         }
 
@@ -136,11 +157,22 @@ public class CriterionService {
 
     /**
      * Deletes a specific criteria item by ID.
+     *
      * @param id The ID of the criteria item to delete.
      * @return True if the criteria item was deleted, false otherwise.
      */
+   // @CheckPublishedRelation(targetClass = Criterion.class,permittedStatus = false)
     @Transactional
-    public boolean delete(String id) {
+   public boolean delete(String id) {
+
+        if(criterionActorRepository.existCriterionInStatus(id, Boolean.TRUE)){
+            throw new ForbiddenException("No action permitted , criterion exists in a published motivation");
+        }
+
+        if (principleCriterionRepository.existsByCriterion(id)) {
+
+            throw new ForbiddenException("This Criterion cannot be deleted because it is linked to a Motivation.");
+        }
 
         return criteriaRepository.deleteById(id);
     }
@@ -153,10 +185,11 @@ public class CriterionService {
 
     /**
      * Retrieves a page of criteria items.
+     *
      * @param motivationId The motivation id
-     * @param page    The index of the page to retrieve (starting from 0).
-     * @param size    The maximum number of criteria items to include in a page.
-     * @param uriInfo The Uri Info.
+     * @param page         The index of the page to retrieve (starting from 0).
+     * @param size         The maximum number of criteria items to include in a page.
+     * @param uriInfo      The Uri Info.
      * @return A PageResource containing the criteria items in the requested page.
      */
     public PageResource<PrincipleCriterionResponse> listCriteriaByMotivation(String motivationId, int page, int size, UriInfo uriInfo) {
@@ -170,13 +203,13 @@ public class CriterionService {
      * Retrieves a detailed Motivation Criterion, including the Metric and associated Metric Tests, by its Motivation ID and Criterion ID.
      *
      * @param motivationId the ID of the Motivation.
-     * @param criterionId the ID of the Motivation Criterion to retrieve.
+     * @param criterionId  the ID of the Motivation Criterion to retrieve.
      * @return the Motivation Criterion if found.
      * @throws NotFoundException if the Motivation Criterion is not found.
      */
     public DetailedCriterionDto getMotivationCriterion(String motivationId, String criterionId) {
 
-        criterionMetricRepository.fetchCriterionMetricByMotivationAndCriterion(motivationId, criterionId).orElseThrow(()-> new NotFoundException("Not Found Criterion."));
+        criterionMetricRepository.fetchCriterionMetricByMotivationAndCriterion(motivationId, criterionId).orElseThrow(() -> new NotFoundException("Not Found Criterion."));
 
         var cri = criteriaRepository.fetchMotivationCriterion(motivationId, criterionId);
 
@@ -185,7 +218,7 @@ public class CriterionService {
 
         for (var row : cri) {
 
-            Node mtrNode = mtrMap.computeIfAbsent(row.getMTR(), k -> new MetricNode(k, row.getLabelMetric().trim(), row.getLabelBenchmarkType(), Double.parseDouble(row.getValueBenchmark())));
+            Node mtrNode = mtrMap.computeIfAbsent(row.getMTR(), k -> new MetricNode(k, row.getLabelMetric().trim(), row.getLabelBenchmarkType(), Double.parseDouble(row.getValueBenchmark()), row.getLabelAlgorithmType(), row.getLabelTypeMetric()));
             Node testNode = testMap.computeIfAbsent(row.getTES(), k -> new TestNode(k, row.getLabelTest().trim(), row.getDescTest().trim(), row.getLabelTestMethod().trim(), row.getTestQuestion(), row.getTestParams(), row.getToolTip()));
 
             if (!mtrNode.getChildren().contains(testNode)) {
@@ -201,4 +234,67 @@ public class CriterionService {
 
         return detailed;
     }
+
+
+    /**
+     * This method takes a Criterion entity, converts it to a CriterionResponse, retrieves and maps
+     * any associated motivations, and then sets the motivations in the response.
+     *
+     * @param criterion The Criterion entity to be converted and enhanced.
+     * @return A CriterionResponse with associated motivations.
+     */
+    private CriterionResponse criterionResponseWithMotivations(Criterion criterion) {
+
+        var criterionResponse = CriteriaMapper.INSTANCE.criteriaToDto(criterion);
+
+        var motivations = principleCriterionRepository.getMotivationIdsByCriterion(criterion.getId());
+        var motivationResponses = motivations.stream()
+                .map(MotivationMapper.INSTANCE::mapPartialMotivation)
+                .collect(Collectors.toList());
+
+        criterionResponse.setMotivations(motivationResponses);
+
+        var metrics = criterionMetricRepository.findMetricsByCriterionId(criterion.getId());
+
+        Map<String, MetricNode> metricNodeMap = new HashMap<>();
+
+        for (var row : metrics) {
+            var metricNode = metricNodeMap.computeIfAbsent(
+                    row.getMTR(),
+                    key -> new MetricNode(
+                            row.getMTR(),
+                            row.getLabelMetric().trim(),
+                            row.getLabelBenchmarkType(),
+                            Double.parseDouble(row.getValueBenchmark()),
+                            row.getLabelAlgorithmType(),
+                            row.getLabelTypeMetric()
+                    )
+            );
+
+            var uniqueTestIds = metricNode.getChildren().stream()
+                    .map(child -> ((TestNode) child).getId())
+                    .collect(Collectors.toSet());
+
+            if (!uniqueTestIds.contains(row.getTES())) {
+
+                var transformedParams = TestParamsTransformer.transformTestParams(row.getTestParams());
+
+                var testNode = new TestNode(
+                        row.getTES(),
+                        row.getLabelTest().trim(),
+                        row.getDescTest().trim(),
+                        row.getLabelTestMethod(),
+                        row.getTestQuestion(),
+                        transformedParams,
+                        row.getToolTip()
+                );
+                metricNode.addChild(testNode);
+            }
+        }
+
+        criterionResponse.setMetrics(new ArrayList<>(metricNodeMap.values()));
+
+        return criterionResponse;
+    }
+
 }
