+++ service/src/main/java/org/grnet/cat/services/registry/TestDefinitionService.java
@@ -4,14 +4,15 @@ import io.quarkus.hibernate.orm.panache.Panache;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.UriInfo;
 import org.grnet.cat.dtos.pagination.PageResource;
-import org.grnet.cat.dtos.registry.test.TestDefinitionRequestDto;
-import org.grnet.cat.dtos.registry.test.TestDefinitionResponseDto;
-import org.grnet.cat.dtos.registry.test.TestDefinitionUpdateDto;
+import org.grnet.cat.dtos.registry.test.*;
+import org.grnet.cat.entities.registry.TestDefinition;
 import org.grnet.cat.entities.registry.TestMethod;
 import org.grnet.cat.mappers.registry.TestDefinitionMapper;
+import org.grnet.cat.repositories.registry.MetricTestRepository;
 import org.grnet.cat.repositories.registry.TestDefinitionRepository;
 import org.grnet.cat.repositories.registry.TestMethodRepository;
 import org.jboss.logging.Logger;
@@ -26,6 +27,10 @@ public class TestDefinitionService {
     
     @Inject
     TestMethodRepository testMethodRepository;
+
+    @Inject
+    MetricTestRepository metricTestRepository;
+
     private static final Logger LOG = Logger.getLogger(TestDefinitionService.class);
 
     /**
@@ -44,21 +49,43 @@ public class TestDefinitionService {
     /**
      * Creates a new TestDefinition item.
      *
-     * @param userId The user creating the TestDefinition.
-     * @param testDefinitionRequestDto The TestDefinition request data.
+     * @param userId The user creating the TestAndTestDefinitionRequest.
+     * @param request The TestAndTestDefinitionRequest request data.
      * @return The created TestDefinition DTO.
      */
     @Transactional
-    public TestDefinitionResponseDto createTestDefinition(String userId, TestDefinitionRequestDto testDefinitionRequestDto) {
+    public TestDefinitionResponseDto createTestDefinition(String userId, TestDefinitionRequestDto request) {
 
-        var testDefinition = TestDefinitionMapper.INSTANCE.testDefinitionToEntity(testDefinitionRequestDto);
+        var testDefinition = TestDefinitionMapper.INSTANCE.testDefinitionToEntity(request);
 
         testDefinition.setPopulatedBy(userId);
-        testDefinition.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, testDefinitionRequestDto.testMethodId));
+        testDefinition.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, request.testMethodId));
 
         testDefinitionRepository.persist(testDefinition);
 
         return TestDefinitionMapper.INSTANCE.testDefinitionToDto(testDefinition);
+    }
+
+
+    /**
+     * Creates a new TestDefinition item.
+     *
+     * @param userId The user creating the TestAndTestDefinitionRequest.
+     * @param request The TestAndTestDefinitionRequest request data.
+     * @return The created TestDefinition DTO.
+     */
+    @Transactional
+    public TestDefinition createTestDefinition(String userId, TestDefinitionRequestDto request, String lodTest) {
+
+        var testDefinition = TestDefinitionMapper.INSTANCE.testDefinitionToEntity(request);
+
+        testDefinition.setPopulatedBy(userId);
+        testDefinition.setLodTES(lodTest);
+        testDefinition.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, request.testMethodId));
+
+        testDefinitionRepository.persist(testDefinition);
+
+        return testDefinition;
     }
 
     /**
@@ -71,6 +98,10 @@ public class TestDefinitionService {
      */
     @Transactional
     public TestDefinitionResponseDto updateTestDefinition(String id, String userId, TestDefinitionUpdateDto request) {
+
+        if (metricTestRepository.existTestDefinitionInStatus(id, Boolean.TRUE)) {
+            throw new ForbiddenException("No action permitted, test definition exists in a published motivation");
+        }
 
         var testDefinition = testDefinitionRepository.findById(id);
 
@@ -92,7 +123,9 @@ public class TestDefinitionService {
      */
     @Transactional
     public boolean deleteTestDefinition(String id) {
-
+        if (metricTestRepository.existTestDefinitionInStatus(id, Boolean.TRUE)) {
+            throw new ForbiddenException("No action permitted, test definition exists in a published motivation");
+        }
         return testDefinitionRepository.deleteById(id);
     }
